move AbstractComputationService to common
